onSensorChanged() now checks if there's a connection before sending commands
Created the method getConnectionState() in the Car class that returns bConnected. If bConnected is false, onSensorChanged doesn't send commands to the car, if bConnected is true, it does.

Also changed the calculation of the float speed, so that now you don't have to hold the device flat but you can hold it in a more natural way